[CS-B-025] Перенесла из ttmanager в админку файлы, связанные с RefClassDurations, а также GroupsController и GroupsService.
