+++ server/admin/src/main/java/ru/cfuv/cfuvscheduling/admin/bom/RefClassDurationsBom.java
@@ -1,4 +1,4 @@
-package ru.cfuv.cfuvscheduling.ttmanager.bom;
+package ru.cfuv.cfuvscheduling.admin.bom;
 
 import lombok.Getter;
 import lombok.Setter;
