+++ server/admin/src/main/java/ru/cfuv/cfuvscheduling/admin/service/GroupsService.java
@@ -1,11 +1,13 @@
-package ru.cfuv.cfuvscheduling.ttmanager.service;
+package ru.cfuv.cfuvscheduling.admin.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import ru.cfuv.cfuvscheduling.commons.bom.GroupsBom;
 import ru.cfuv.cfuvscheduling.commons.converter.GroupsConverter;
 import ru.cfuv.cfuvscheduling.commons.dao.GroupsDao;
 import ru.cfuv.cfuvscheduling.commons.dao.dto.GroupsDto;
+import ru.cfuv.cfuvscheduling.commons.exception.AlreadyExistsException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,6 +17,20 @@ public class GroupsService {
 
     @Autowired
     private GroupsDao groupsDao;
+
+    public GroupsBom addNewGroup(GroupsBom group) {
+        try {
+            group.setId(null);
+            GroupsDto groupDto = new GroupsDto();
+            GroupsConverter converter = new GroupsConverter();
+            converter.toDto(group, groupDto);
+            groupsDao.save(groupDto);
+            converter.fromDto(groupDto, group);
+            return group;
+        } catch (DataIntegrityViolationException e) {
+            throw new AlreadyExistsException("Group with name %s already exist".formatted(group.getName()));
+        }
+    }
 
     public List<GroupsBom> findAll() {
         List<GroupsDto> dtos = groupsDao.findAll();
@@ -28,5 +44,4 @@ public class GroupsService {
 
         return boms;
     }
-
 }
