--- server/tt-manager/src/main/java/ru/cfuv/cfuvscheduling/ttmanager/controller/GroupsController.java
@@ -1,23 +1,0 @@
-package ru.cfuv.cfuvscheduling.ttmanager.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import ru.cfuv.cfuvscheduling.commons.bom.GroupsBom;
-import ru.cfuv.cfuvscheduling.ttmanager.service.GroupsService;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/group")
-public class GroupsController {
-
-    @Autowired
-    private GroupsService groupsService;
-
-    @GetMapping("/findAll")
-    public List<GroupsBom> findAllGroups() {
-        return groupsService.findAll();
-    }
-}
